+++ telegram_bot/redis.py
@@ -2,4 +2,4 @@ from redis import Redis
 
 from telegram_bot.settings import conf
 
-redis_conn = Redis.from_url(conf.REDIS_URL)
+redis_conn = Redis.from_url(conf['REDIS_URL'])