+++ telegram_bot/signals.py
@@ -10,7 +10,7 @@ def autodiscover_tg_routers():
     for app_config in apps.get_app_configs():
         try:
             # Name of the module containing 'tg_router', e.g. 'myapp.tg_router'
-            routers_module_name = f"{app_config.name}.{conf.MODULE_NAME}"
+            routers_module_name = f"{app_config.name}.{conf['MODULE_NAME']}"
             import_module(routers_module_name)
 
         except ImportError:  # tg_router not exists in that app